apps/worker/pyproject.toml
[project]
name = "alttext-worker"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic>=2",
  "httpx",
  "pillow"
]

